# Preprocessing the frames more faster