## Time results for the same video.

### Before changes
 - vtt - 36.88 secs
 - json - 33.95 secs

### After changes
 - 27.03 secs
 - 16.68 secs